Make two fields private, move a method
